[kk1219] Changed read_in_prog to use uint8

+++ src/assemble_src/headers/assemble.h
@@ -5,6 +5,6 @@
 
 void save_instruction(uint32_t *assembled_program, uint16_t address ,uint32_t data);
 
-char **read_in_prog(char *filename, int *num_instr);
+char **read_in_prog(char *filename, uint8_t *num_instr);
 
 #endif
